top and bottom bars complete.
