+++ DietScoopApp/app/src/main/res/drawable/ic_baseline_bento_24.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#000000"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M16,11V5h4c1.1,0 2,0.9 2,2v4H16zM20,19c1.1,0 2,-0.9 2,-2v-4h-6v6H20zM14,5v14H4c-1.1,0 -2,-0.9 -2,-2V7c0,-1.1 0.9,-2 2,-2H14zM9.5,12c0,-0.83 -0.67,-1.5 -1.5,-1.5S6.5,11.17 6.5,12s0.67,1.5 1.5,1.5S9.5,12.83 9.5,12z"/>
+</vector>
