+++ DietScoopApp/app/src/main/res/drawable/ic_baseline_shopping_bag_24.xml
@@ -1,0 +1,5 @@
+<vector android:height="24dp" android:tint="#000000"
+    android:viewportHeight="24" android:viewportWidth="24"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="@android:color/white" android:pathData="M18,6h-2c0,-2.21 -1.79,-4 -4,-4S8,3.79 8,6H6C4.9,6 4,6.9 4,8v12c0,1.1 0.9,2 2,2h12c1.1,0 2,-0.9 2,-2V8C20,6.9 19.1,6 18,6zM10,10c0,0.55 -0.45,1 -1,1s-1,-0.45 -1,-1V8h2V10zM12,4c1.1,0 2,0.9 2,2h-4C10,4.9 10.9,4 12,4zM16,10c0,0.55 -0.45,1 -1,1s-1,-0.45 -1,-1V8h2V10z"/>
+</vector>
