+++ DietScoopApp/app/src/main/res/drawable/ic_egg_black_24dp.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M12,3C8.5,3 5,9.33 5,14c0,3.87 3.13,7 7,7s7,-3.13 7,-7C19,9.33 15.5,3 12,3zM13,18c-3,0 -5,-1.99 -5,-5c0,-0.55 0.45,-1 1,-1s1,0.45 1,1c0,2.92 2.42,3 3,3c0.55,0 1,0.45 1,1S13.55,18 13,18z"
+      android:fillColor="#000000"/>
+</vector>
