+++ DietScoopApp/app/src/main/res/menu/bottom_nav_menu.xml
@@ -3,22 +3,22 @@
     <item
         android:id="@+id/shopping"
         android:title="Shopping"
-        android:icon="@drawable/ic_launcher_background"
+        android:icon="@drawable/ic_baseline_shopping_bag_24"
         android:enabled="true"/>
     <item
         android:id="@+id/meals"
         android:title="Meals"
-        android:icon="@drawable/ic_launcher_background"
+        android:icon="@drawable/ic_baseline_bento_24"
         android:enabled="true"/>
     <item
         android:id="@+id/recipes"
         android:title="Recipes"
-        android:icon="@drawable/ic_launcher_background"
+        android:icon="@drawable/ic_baseline_menu_book_24"
         android:enabled="true"/>
     <item
         android:id="@+id/ingredients"
         android:title="Ingredients"
-        android:icon="@drawable/ic_launcher_background"
+        android:icon="@drawable/ic_egg_black_24dp"
         android:enabled="true"/>
 
 </menu>
